chat auto scroll fixed
