+++ chat/src/pages/Chat/ChatBody.jsx
@@ -1,8 +1,16 @@
-import React from "react";
+import React, { useEffect, useRef } from "react";
 import styles from "./ChatBody.module.scss";
 import { MessageItem } from "../../components/MessageItem/MessageItem.jsx";
 
-export const ChatBody = ({ messages, chatBodyRef }) => {
+export const ChatBody = ({ messages }) => {
+  const chatBodyRef = useRef(null);
+
+  useEffect(() => {
+      if (chatBodyRef.current) {
+        chatBodyRef.current.scrollIntoView({ behavior: "smooth" });
+      }
+    }, [messages]);
+
   return (
     <div className={styles["chat-body"]}>
       {messages.map((message) => (
