+++ chat/src/pages/Chat/ChatBody.module.scss
@@ -1,5 +1,5 @@
 .chat-body {
-	height: calc(100vh - 100px);
+	height: calc(100vh - 137.8px);
 	overflow-y: auto;
 	overflow-x: hidden;
 	padding: 10px;
